requirements.txt
Flask==1.0.3
Jinja2==2.10.1
firebase_admin==4.5.3
Pyrebase4==4.5.0
pandas==1.2.4